#62 Reorganized root components to better show app structure
