+++ src/TripPlanner.App/ClientApp/src/root/App.tsx
@@ -5,11 +5,13 @@ import {
 import { LocalizationProvider } from '@mui/lab';
 import DateAdapter from '@mui/lab/AdapterDayjs';
 import { useRecoilValue } from 'recoil';
-import ToasterNotificationsComponent from './components/ToasterNotifications';
-import DebugStateObserver from './State/DebugStateObserver';
-import { isAppLoadedState } from './State/AppState';
-import App from './App';
+import ToasterNotificationsComponent from '../components/ToasterNotifications';
+import DebugStateObserver from '../State/DebugStateObserver';
+import { isAppLoadedState } from '../State/AppState';
+import AppContent from './AppContent';
 import AppLoader from './AppLoader';
+import LocationActionLoadingIndicator from '../components/LocationActionLoadingIndicator';
+import ModalContainer from '../components/modals/ModalContainer';
 
 declare module '@mui/styles/defaultTheme' {
   // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -19,7 +21,7 @@ declare module '@mui/styles/defaultTheme' {
 let theme = createTheme();
 theme = responsiveFontSizes(theme);
 
-function AppContent() {
+function App() {
   const isAppLoaded = useRecoilValue(isAppLoadedState);
 
   return (
@@ -28,8 +30,10 @@ function AppContent() {
       <ToasterNotificationsComponent />
       <StyledEngineProvider injectFirst>
         <ThemeProvider theme={theme}>
+          <LocationActionLoadingIndicator />
+          <ModalContainer />
           <LocalizationProvider dateAdapter={DateAdapter}>
-            {isAppLoaded ? <App /> : <AppLoader />}
+            {isAppLoaded ? <AppContent /> : <AppLoader />}
           </LocalizationProvider>
         </ThemeProvider>
       </StyledEngineProvider>
@@ -38,4 +42,4 @@ function AppContent() {
   );
 }
 
-export default AppContent;
+export default App;
